🐛Fix: 자잘한 api 수정(section 페이지 action item 추가, 따로 컴포넌트 설정)
