--- src/api/__mock__/retrospectiveTemplate.ts
@@ -1,9 +1,0 @@
-import { RetrospectivesTemplateResponse } from '../@types/RetrospectiveTemplates';
-
-export const mockRetrospectiveTemplate: RetrospectivesTemplateResponse[] = [
-  // { id: 1, name: 'hee' },
-  // {
-  //   id: 2,
-  //   name: 'jung',
-  // },
-];
