+++ src/api/services/Retrospectives.ts
@@ -9,7 +9,6 @@ import {
   RetrospectivesClient,
 } from '../@types/Retrospectives';
 import axiosInstance from '../axiosConfig';
-import { mswInstance } from '../client';
 
 const ROUTE = 'retrospectives';
 
@@ -32,7 +31,7 @@ export const RetrospectiveService: RetrospectivesClient = {
   },
   get: async (request: GetRetrospectiveRequest): Promise<GetRetrospectiveData> => {
     try {
-      const response = await mswInstance.get(`${ROUTE}/`, {
+      const response = await axiosInstance.get(`${ROUTE}/`, {
         params: request,
       });
       return response.data;
@@ -43,12 +42,13 @@ export const RetrospectiveService: RetrospectivesClient = {
 
   delete: async ({ retrospectiveId }: DeleteRetrospectiveRequest): Promise<void> => {
     try {
-      const response = await mswInstance.delete(`${ROUTE}/${retrospectiveId}`);
+      const response = await axiosInstance.delete(`${ROUTE}/${retrospectiveId}`);
       return response.data;
     } catch (error) {
       throw new Error(error as string);
     }
   },
+
   put: async ({ retrospectiveId }, ...request) => {
     try {
       const response = await axiosInstance.put(`${ROUTE}/${retrospectiveId}`, request);
@@ -57,7 +57,8 @@ export const RetrospectiveService: RetrospectivesClient = {
       throw new Error(error as string);
     }
   },
+
   patch: async (retrospectiveId, ...request) => {
-    return await mswInstance.patch(`${ROUTE}/${retrospectiveId}/bookmark`, request);
+    return await axiosInstance.patch(`${ROUTE}/${retrospectiveId}/bookmark`, request);
   },
 };
