+++ src/styles/RetroList/RetroListPage.style.ts
@@ -27,6 +27,8 @@ export const SortButtonContainer = styled.div`
 export const Box = styled.div`
   padding-left: 40px;
   padding-right: 40px;
+  margin: 0 auto;
+  width: 1000px;
 `;
 
 export const ControBarContainer = styled.div`
